fixed 5 issues
fixed 5 issues
Remove color_scheme from Attributes in NFT slide over because these are parameters that are read-only once NFT is minted and cannot be changed. Move this to Profile section.

Background for the Input and Select boxes in NFTSlideOverLeft.jsx are off. Sometimes transparent, sometimes pure white. .

In the UI, text aria-label=… appears in the color customisation section in the buttons.

For some reason, once I set the Primary and Secondary colours and saved, every button and input box went invisible/black until I changed the Neutrals slider to more saturated and then saved again. Then the UI controls appeared again.

As mentioned before, the 6 utility colours have to be editable so the user can change them. Need to be able to click their button and for the color picker to come up (like the primary and secondary). Keep their default colours you have there as the starting point.